backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
numpy>=1.26.3
python-multipart>=0.0.6
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
sqlalchemy>=2.0.0
cryptography>=41.0.0
replicate>=0.22.0
scipy>=1.11.0
requests>=2.31.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
